+++ src/renderer/src/components/common/DiffViewer/index.tsx
@@ -1,1 +1,2 @@
 export * from './DiffViewer';
+export * from './UDiffViewer';